+++ lib/fuery_core.dart
@@ -9,6 +9,11 @@ export 'src/query_options.dart';
 export 'src/query_state.dart';
 export 'src/query_result.dart';
 
+// infinite query
+export 'src/infinite_query.dart' show InfiniteQuery;
+export 'src/infinite_data.dart';
+export 'src/infinite_query_result.dart';
+
 // mutation
 export 'src/mutation.dart' show Mutation;
 export 'src/mutation_options.dart';