+++ lib/src/error/fuery_exception.dart
@@ -1,5 +1,6 @@
 part 'invalid_timestamp.dart';
 part 'query_state_type.dart';
+part 'selector_not_found.dart';
 
 sealed class FueryException {
   const FueryException({this.message = ''});