+++ lib/src/error/selector_not_found.dart
@@ -1,0 +1,5 @@
+part of 'fuery_exception.dart';
+
+class SelectorNotFound extends FueryException {
+  SelectorNotFound({super.message = 'InfiniteQuery has no data selector.'});
+}