+++ lib/src/mutation.dart
@@ -8,10 +8,11 @@ import 'package:fuery_core/src/mutation_options.dart';
 import 'package:fuery_core/src/mutation_result.dart';
 import 'package:fuery_core/src/mutation_state.dart';
 
-part 'mutation_with_args.dart';
-part 'mutation_without_args.dart';
+part 'mutation_with_params.dart';
+part 'mutation_without_params.dart';
 
-sealed class Mutation<Args, Data, Err> extends MutationBase<Args, Data, Err, MutationState<Data, Err>> {
+sealed class Mutation<Params, Data, Err>
+    extends MutationBase<Params, Data, Err, MutationState<Data, Err>> {
   Mutation._({
     MutationKey? mutationKey,
     super.options,
@@ -41,22 +42,25 @@ sealed class Mutation<Args, Data, Err> extends MutationBase<Args, Data, Err, Mut
     }
   }
 
-  static MutationResult<Data, Err, MutationSyncFn<Args, Data, Err>, MutationAsyncFn<Args, Data, Err>> args<Args, Data, Err>({
-    required MutationAsyncFn<Args, Data, Err> mutationFn,
+  static MutationResult<Data, Err, MutationSyncFn<Params, Data, Err>,
+      MutationAsyncFn<Params, Data, Err>> use<Params, Data, Err>({
+    required MutationAsyncFn<Params, Data, Err> mutationFn,
     MutationKey? mutationKey,
     int? gcTime,
-    MutationMutateCallback<Args>? onMutate,
-    MutationSuccessCallback<Args, Data>? onSuccess,
-    MutationErrorCallback<Args, Err>? onError,
+    MutationMutateCallback<Params>? onMutate,
+    MutationSuccessCallback<Params, Data>? onSuccess,
+    MutationErrorCallback<Params, Err>? onError,
   }) {
     try {
-      final MutationWithArgs<Args, Data, Err> mutation = _getCachedMutation<MutationWithArgs<Args, Data, Err>>(
+      final MutationWithParams<Params, Data, Err> mutation =
+          _getCachedMutation<MutationWithParams<Params, Data, Err>>(
         mutationKey: mutationKey,
         orElse: () {
-          final MutationWithArgs<Args, Data, Err> mutation = MutationWithArgs<Args, Data, Err>(
+          final MutationWithParams<Params, Data, Err> mutation =
+              MutationWithParams<Params, Data, Err>(
             mutationKey: mutationKey,
             mutationFn: mutationFn,
-            options: MutationOptions<Args, Data, Err>(
+            options: MutationOptions<Params, Data, Err>(
               gcTime: gcTime,
               onMutate: onMutate,
               onSuccess: onSuccess,
@@ -64,11 +68,12 @@ sealed class Mutation<Args, Data, Err> extends MutationBase<Args, Data, Err, Mut
             ),
           );
 
-          if (mutationKey != null) Fuery.instance.addMutation(mutationKey, mutation);
+          if (mutationKey != null)
+            Fuery.instance.addMutation(mutationKey, mutation);
 
           return mutation;
         },
-      ) as MutationWithArgs<Args, Data, Err>;
+      ) as MutationWithParams<Params, Data, Err>;
 
       return MutationResult(
         data: mutation.stream,
@@ -80,19 +85,22 @@ sealed class Mutation<Args, Data, Err> extends MutationBase<Args, Data, Err, Mut
     }
   }
 
-  static MutationResult<Data, Err, MutationNoArgsSyncFn<Data, Err>, MutationNoArgsAsyncFn<Data, Err>> noArgs<Data, Err>({
-    required MutationNoArgsAsyncFn<Data, Err> mutationFn,
+  static MutationResult<Data, Err, MutationNoParamSyncFn,
+      MutationNoParamAsyncFn<Data, Err>> noParam<Data, Err>({
+    required MutationNoParamAsyncFn<Data, Err> mutationFn,
     MutationKey? mutationKey,
     int? gcTime,
-    MutationMutateCallbackWithoutArgs? onMutate,
-    MutationSuccessCallbackWithoutArgs<Data>? onSuccess,
-    MutationErrorCallbackWithoutArgs<Err>? onError,
+    MutationNoParamMutateCallback? onMutate,
+    MutationNoParamSuccessCallback<Data>? onSuccess,
+    MutationNoParamErrorCallback<Err>? onError,
   }) {
     try {
-      final MutationNoArgs<Data, Err> mutation = _getCachedMutation<MutationNoArgs<Data, Err>>(
+      final MutationNoParams<Data, Err> mutation =
+          _getCachedMutation<MutationNoParams<Data, Err>>(
         mutationKey: mutationKey,
         orElse: () {
-          final MutationNoArgs<Data, Err> mutation = MutationNoArgs<Data, Err>(
+          final MutationNoParams<Data, Err> mutation =
+              MutationNoParams<Data, Err>(
             mutationKey: mutationKey,
             mutationFn: mutationFn,
             options: MutationOptions(
@@ -102,11 +110,12 @@ sealed class Mutation<Args, Data, Err> extends MutationBase<Args, Data, Err, Mut
             ),
           );
 
-          if (mutationKey != null) Fuery.instance.addMutation(mutationKey, mutation);
+          if (mutationKey != null)
+            Fuery.instance.addMutation(mutationKey, mutation);
 
           return mutation;
         },
-      ) as MutationNoArgs<Data, Err>;
+      ) as MutationNoParams<Data, Err>;
 
       return MutationResult(
         data: mutation.stream,