+++ lib/src/mutation_options.dart
@@ -1,42 +1,42 @@
 import 'package:fuery_core/src/base/typedefs.dart';
 
-class MutationOptions<Args, Data, Err> {
+class MutationOptions<Params, Data, Err> {
   MutationOptions({
-    Args? arguments,
+    Params? params,
     int? gcTime,
-    MutationMutateCallback<Args>? onMutate,
-    MutationSuccessCallback<Args, Data>? onSuccess,
-    MutationErrorCallback<Args, Err>? onError,
-  })  : _arguments = arguments,
+    MutationMutateCallback<Params>? onMutate,
+    MutationSuccessCallback<Params, Data>? onSuccess,
+    MutationErrorCallback<Params, Err>? onError,
+  })  : _params = params,
         _gcTime = gcTime ?? 0,
         _onMutate = onMutate,
         _onSuccess = onSuccess,
         _onError = onError;
 
-  final Args? _arguments;
-  Args? get arguments => _arguments;
+  final Params? _params;
+  Params? get params => _params;
 
   final int _gcTime;
   int get gcTime => _gcTime;
 
-  final MutationMutateCallback<Args>? _onMutate;
-  MutationMutateCallback<Args>? get onMutate => _onMutate;
+  final MutationMutateCallback<Params>? _onMutate;
+  MutationMutateCallback<Params>? get onMutate => _onMutate;
 
-  final MutationSuccessCallback<Args, Data>? _onSuccess;
-  MutationSuccessCallback<Args, Data>? get onSuccess => _onSuccess;
+  final MutationSuccessCallback<Params, Data>? _onSuccess;
+  MutationSuccessCallback<Params, Data>? get onSuccess => _onSuccess;
 
-  final MutationErrorCallback<Args, Err>? _onError;
-  MutationErrorCallback<Args, Err>? get onError => _onError;
+  final MutationErrorCallback<Params, Err>? _onError;
+  MutationErrorCallback<Params, Err>? get onError => _onError;
 
-  MutationOptions<Args, Data, Err> copyWith({
-    ValueGetter<Args>? arguments,
+  MutationOptions<Params, Data, Err> copyWith({
+    ValueGetter<Params>? params,
     ValueGetter<int>? gcTime,
-    ValueGetter<MutationMutateCallback<Args>>? onMutate,
-    ValueGetter<MutationSuccessCallback<Args, Data>>? onSuccess,
-    ValueGetter<MutationErrorCallback<Args, Err>>? onError,
+    ValueGetter<MutationMutateCallback<Params>>? onMutate,
+    ValueGetter<MutationSuccessCallback<Params, Data>>? onSuccess,
+    ValueGetter<MutationErrorCallback<Params, Err>>? onError,
   }) {
-    return MutationOptions<Args, Data, Err>(
-      arguments: arguments != null ? arguments() : this.arguments,
+    return MutationOptions<Params, Data, Err>(
+      params: params != null ? params() : this.params,
       onMutate: onMutate != null ? onMutate() : this.onMutate,
       onSuccess: onSuccess != null ? onSuccess() : this.onSuccess,
       onError: onError != null ? onError() : this.onError,
@@ -47,8 +47,8 @@ class MutationOptions<Args, Data, Err> {
   bool operator ==(Object other) {
     if (identical(this, other)) return true;
 
-    return other is MutationOptions<Args, Data, Err> &&
-        other._arguments == _arguments &&
+    return other is MutationOptions<Params, Data, Err> &&
+        other._params == _params &&
         other._gcTime == _gcTime &&
         other._onMutate == _onMutate &&
         other._onSuccess == _onSuccess &&
@@ -57,6 +57,10 @@ class MutationOptions<Args, Data, Err> {
 
   @override
   int get hashCode {
-    return _arguments.hashCode ^ _gcTime.hashCode ^ _onMutate.hashCode ^ _onSuccess.hashCode ^ _onError.hashCode;
+    return _params.hashCode ^
+        _gcTime.hashCode ^
+        _onMutate.hashCode ^
+        _onSuccess.hashCode ^
+        _onError.hashCode;
   }
 }