+++ lib/src/query_result.dart
@@ -1,9 +1,10 @@
 import 'package:fuery_core/src/base/query_result.dart';
 import 'package:fuery_core/src/query_state.dart';
 
-class QueryResult<Data, Err> extends QueryResultBase<Data, Err, QueryState<Data, Err>> {
+class QueryResult<Data, Err>
+    extends QueryResultBase<Data, Err, QueryState<Data, Err>> {
   QueryResult({
-    required super.data,
+    required super.stream,
     required super.refetch,
     required super.updateOptions,
   });