+++ CHANGELOG.md
@@ -1,2 +1,9 @@
 ## 0.0.1
-* Fuery core pre release.
+- chore: Fuery core pre release.
+
+## 0.0.2
+- docs: update dependencies and license.
+
+## 0.0.3
+- feat: InfiniteQuery
+- docs: InfiniteQuery usage